pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Michael P", email = "mipeng@umich.edu"},
]
dependencies = [
    "discord.py @ git+https://github.com/Rapptz/discord.py",
    "pony>=0.7.16",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
lint = [
    "black>=22.6.0",
]
